Fix : Nav 컴포넌트 수정
