+++ src/components/Nav/Nav.jsx
@@ -1,10 +1,17 @@
-import React from 'react'
-import * as S from './NavStyle'
-import Profile from './Profile'
+import React from "react";
+import * as S from "./NavStyle";
 
-import logoImg from '../../assets/logo.png'
-import { Button } from '../Button/ButtonStyle'
+import logoImg from "../../assets/logo.png";
+import { Button } from "../Button/ButtonStyle";
 
+function Profile({ email, profileImage }) {
+  return (
+    <S.Profile>
+      <S.ProfileImg src={profileImage} alt="프로필 사진" />
+      <S.ProfileEmail className="flex-center">{email}</S.ProfileEmail>
+    </S.Profile>
+  );
+}
 function Nav({ path, email, profileImageSource }) {
   return (
     <S.NavContainer path={path}>
@@ -18,7 +25,7 @@ function Nav({ path, email, profileImageSource }) {
         )}
       </S.NavContent>
     </S.NavContainer>
-  )
+  );
 }
 
-export default Nav
+export default Nav;
